Split PrepareTransparent off Inliner
